+++ tt_metal/hw/ckernels/blackhole/metal/llk_api/llk_sfpu/llk_math_eltwise_unary_sfpu_comp_int.h
@@ -6,7 +6,6 @@
 
 #include "llk_math_eltwise_unary_sfpu_init.h"
 #include "llk_math_eltwise_unary_sfpu_params.h"
-#include "ckernel_sfpu_unary_comp.h"
 #include "ckernel_sfpu_comp_int.h"
 
 namespace ckernel {
@@ -34,17 +33,6 @@ inline void llk_math_eltwise_unary_sfpu_unary_ne_int(
     uint dst_index, uint param0, int vector_mode = (int)VectorMode::RC) {
     llk_math_eltwise_unary_sfpu_params<APPROXIMATE>(
         ckernel::sfpu::calculate_unary_ne_int32<APPROXIMATE>, dst_index, vector_mode, param0);
-}
-
-template <bool APPROXIMATE>
-inline void llk_math_eltwise_unary_sfpu_unary_ne_init() {
-    llk_math_eltwise_unary_sfpu_init<SfpuType::unary_ne, APPROXIMATE>();
-}
-
-template <bool APPROXIMATE>
-inline void llk_math_eltwise_unary_sfpu_unary_ne(uint dst_index, uint param0, int vector_mode = (int)VectorMode::RC) {
-    llk_math_eltwise_unary_sfpu_params<APPROXIMATE>(
-        ckernel::sfpu::calculate_unary_ne<APPROXIMATE>, dst_index, vector_mode, param0);
 }
 
 }  // namespace ckernel
