+++ tt_metal/hw/ckernels/wormhole_b0/metal/llk_api/llk_sfpu/llk_math_eltwise_unary_sfpu_unary_comp.h
@@ -12,6 +12,18 @@ namespace ckernel {
 
 // New LLK SFPU APIs
 
+// Unary Not equal
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_unary_ne_init() {
+    llk_math_eltwise_unary_sfpu_init<SfpuType::unary_ne, APPROXIMATE>();
+}
+
+template <bool APPROXIMATE>
+inline void llk_math_eltwise_unary_sfpu_unary_ne(uint dst_index, uint param0, int vector_mode = (int)VectorMode::RC) {
+    llk_math_eltwise_unary_sfpu_params<APPROXIMATE>(
+        ckernel::sfpu::calculate_unary_ne<APPROXIMATE>, dst_index, vector_mode, param0);
+}
+
 // Unary greater than
 template <bool APPROXIMATE>
 inline void llk_math_eltwise_unary_sfpu_unary_gt_init() {
